Tidy up code vastly by removing my flawed idea of having different reminder classes for Themes and Goals.
