+++ .env.dev
@@ -2,4 +2,6 @@
 # I actually *do* have a working Sendmail on port 25 on my Mac, better safe than
 # sorry, eh?
 MAILER_DSN=smtp://localhost:1025
+ADMIN_EMAIL_ADDRESS=fresher-dev@gothick.org.uk
+ADMIN_EMAIL_NAME="Fresher Development System"
