+++ config/services.yaml
@@ -4,12 +4,21 @@
 # Put parameters here that don't need to change on each machine where the app is deployed
 # https://symfony.com/doc/current/best_practices.html#use-parameters-for-application-configuration
 parameters:
+    env(ADMIN_EMAIL_ADDRESS): 'fresher@gothick.org.uk'
+    env(ADMIN_EMAIL_NAME): 'Fresher'
+
+    app.admin_email_address: '%env(ADMIN_EMAIL_ADDRESS)%'
+    app.admin_email_name: '%env(ADMIN_EMAIL_NAME)%'
 
 services:
     # default configuration for services in *this* file
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
+
+        bind:
+            $adminEmailAddress: '%app.admin_email_address%'
+            $adminEmailName: '%app.admin_email_name%'
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
