+++ src/Controller/ReminderController.php
@@ -3,14 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Goal;
-use App\Entity\GoalReminder;
 use App\Entity\Theme;
 use App\Entity\ThemeReminder;
-use App\Form\GoalReminderType;
 use App\Form\ThemeReminderType;
-use ContainerN9NmnxR\getGoalReminderTypeService;
 use Doctrine\ORM\EntityManagerInterface;
-use Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -19,22 +15,6 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 class ReminderController extends AbstractController
 {
-    /**
-     * @Route("/goal/{goal}/reminder", name="goal_reminder")
-     * @IsGranted("access", subject="goal")
-     */
-    public function goalReminderIndex(
-        Goal $goal
-    ): Response {
-        if ($goal->getTheme() === null) {
-            throw new Exception('Expected every goal to have a theme');
-        }
-        return $this->render('reminder/goal_reminder_index.html.twig', [
-            'goal' => $goal,
-            'theme' => $goal->getTheme()
-        ]);
-    }
-
     /**
      * @Route("/theme/{theme}/reminder", name="theme_reminder")
      * @IsGranted("access", subject="theme")
@@ -47,41 +27,6 @@ class ReminderController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/goal/{goal}/reminder/new", name="goal_reminder_new", methods={"GET", "POST"})
-     * @IsGranted("access", subject="goal")
-     */
-    public function newGoalReminder(
-        Goal $goal,
-        Request $request,
-        EntityManagerInterface $entityManager
-    ): Response {
-        if ($goal->getTheme() === null) {
-            throw new Exception('Expected every goal to have a theme');
-        }
-
-        $goalReminder = new GoalReminder();
-        $goalReminder->setGoal($goal);
-        $form = $this->createForm(GoalReminderType::class, $goalReminder);
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            /** @var GoalReminder $goalReminder */
-            $goalReminder = $form->getData();
-            $entityManager->persist($goalReminder);
-            $entityManager->flush();
-            $this->addFlash('success', "New Goal Reminder added.");
-            return $this->redirectToRoute('goal_show', [
-                'theme' => $goal->getTheme()->getId(),
-                'goal' => $goal->getId()
-            ]);
-        }
-
-        return $this->renderForm('reminder/new.html.twig', [
-            'form' => $form,
-            //'goal' => $goal
-        ]);
-    }
     /**
      * @Route("/theme/{theme}/reminder/new", name="theme_reminder_new", methods={"GET", "POST"})
      * @IsGranted("access", subject="theme")
@@ -112,41 +57,6 @@ class ReminderController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/goal/{goal}/reminder/{reminder}/edit", name="goal_reminder_edit", methods={"GET", "POST"})
-     * @IsGranted("access", subject="goalReminder")
-     */
-    public function editGoalReminder(
-        Goal $goal,
-        GoalReminder $goalReminder,
-        Request $request,
-        EntityManagerInterface $entityManager
-    ): Response {
-        if ($goal->getTheme() === null) {
-            throw new Exception('Expected every goal to have a theme');
-        }
-        $form = $this->createForm(GoalReminderType::class, $goalReminder, [
-            'submit_label' => 'Save'
-        ]);
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            /** @var GoalReminder $goalReminder */
-            $goalReminder = $form->getData();
-            $entityManager->flush();
-            $this->addFlash('success', "Goal Reminder edited.");
-            return $this->redirectToRoute('goal_show', [
-                'theme' => $goal->getTheme()->getId(),
-                'goal' => $goal->getId()
-            ]);
-        }
-
-        return $this->renderForm('reminder/edit_goal_reminder.html.twig', [
-            'form' => $form,
-            'goal' => $goal,
-            'reminder' => $goalReminder
-        ]);
-    }
     /**
      * @Route("/theme/{theme}/reminder/{reminder}/edit", name="theme_reminder_edit", methods={"GET", "POST"})
      * @IsGranted("access", subject="themeReminder")
@@ -179,37 +89,6 @@ class ReminderController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/goal/{goal}/reminder/{reminder}/delete", name="goal_reminder_delete", methods={"DELETE"})
-     * @IsGranted("access", subject="goalReminder")
-     */
-    public function deleteGoalReminder(
-        Goal $goal,
-        GoalReminder $goalReminder,
-        Request $request,
-        EntityManagerInterface $entityManager
-    ): Response {
-        // TODO: security
-        if ($goal->getTheme() === null) {
-            throw new Exception('Expected every goal to have a theme');
-        }
-
-        $submittedToken = (string) $request->request->get('token');
-        if ($this->isCsrfTokenValid('goal_reminder_delete', $submittedToken)) {
-            $entityManager->remove($goalReminder);
-            $entityManager->flush();
-            $this->addFlash('success', "Goal Reminder deleted successfully.");
-            return $this->redirectToRoute('goal_show', [
-                'theme' => $goal->getTheme()->getId(),
-                'goal' => $goal->getId()
-            ]);
-        }
-        $this->addFlash('danger', "Goal Reminder could not be deleted.");
-        return $this->redirectToRoute('goal_show', [
-            'theme' => $goal->getTheme()->getId(),
-            'goal' => $goal->getId()
-        ]);
-    }
     /**
      * @Route("/theme/{theme}/reminder/{reminder}/delete", name="theme_reminder_delete", methods={"DELETE"})
      * @IsGranted("access", subject="themeReminder")
