--- src/Form/GoalReminderType.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\GoalReminder;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class GoalReminderType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('enabled')
-            //->add('goal')
-            ->add('create', SubmitType::class, [
-                'label' => $options['submit_label']
-            ]);
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => GoalReminder::class,
-            'submit_label' => 'Create'
-        ]);
-    }
-}
