+++ src/Repository/ThemeReminderJobRepository.php
@@ -2,25 +2,25 @@
 
 namespace App\Repository;
 
-use App\Entity\Reminder;
+use App\Entity\ThemeReminderJob;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 
 /**
- * @method Reminder|null find($id, $lockMode = null, $lockVersion = null)
- * @method Reminder|null findOneBy(array $criteria, array $orderBy = null)
- * @method Reminder[]    findAll()
- * @method Reminder[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ * @method ThemeReminderJob|null find($id, $lockMode = null, $lockVersion = null)
+ * @method ThemeReminderJob|null findOneBy(array $criteria, array $orderBy = null)
+ * @method ThemeReminderJob[]    findAll()
+ * @method ThemeReminderJob[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class ReminderRepository extends ServiceEntityRepository
+class ThemeReminderJobRepository extends ServiceEntityRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
-        parent::__construct($registry, Reminder::class);
+        parent::__construct($registry, ThemeReminderJob::class);
     }
 
     // /**
-    //  * @return Reminder[] Returns an array of Reminder objects
+    //  * @return ReminderJob[] Returns an array of ReminderJob objects
     //  */
     /*
     public function findByExampleField($value)
@@ -37,7 +37,7 @@ class ReminderRepository extends ServiceEntityRepository
     */
 
     /*
-    public function findOneBySomeField($value): ?Reminder
+    public function findOneBySomeField($value): ?ReminderJob
     {
         return $this->createQueryBuilder('r')
             ->andWhere('r.exampleField = :val')
