--- templates/reminder/edit_goal_reminder.html.twig
@@ -1,18 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Fresher - Edit Goal Reminder{% endblock %}
-
-{% block body %}
-<div class="container">
-    {{ form(form) }}
-    <form
-        method="POST"
-        action="{{ path('goal_reminder_delete', { 'goal': reminder.goal.id, 'reminder': reminder.id }) }}"
-        onsubmit="return confirm('Are you sure you want to delete this reminder?');"
-        >
-        <input type="hidden" name="_method" value="DELETE">
-        <input type="hidden" name="token" value="{{ csrf_token('goal_reminder_delete') }}"/>
-        <button type="submit" name="btn_delete" class="btn btn-danger mt-4"><i class="fas fa-trash-alt" aria-hidden="true"></i> Delete Reminder</button>
-    </form>
-</div>
-{% endblock %}
